rpi5-chatbot/requirements.txt
pygame>=2.5.0
requests>=2.31.0
numpy>=1.24.0
pyaudio>=0.2.13
pyserial>=3.5
psutil>=5.9.0
